Minor improvements

+++ os_x/install_npm_packages.sh
@@ -5,11 +5,18 @@ cd "$(dirname "${BASH_SOURCE}")" && source "utils.sh"
 declare -r -a NPM_PACKAGES=(
     "bower"
     "gulp"
+    "jshint"
 )
 
 # ------------------------------------------------------------------------------
 
 main() {
+
+    # Check if `npm` is installed
+    if [ $(cmd_exists "npm") -eq 1 ]; then
+        print_error "npm is required, please install it!\n"
+        exit 1
+    fi
 
     # Make sure the most recent version of `npm` is installed
     execute "npm install --global npm" "npm (update)"
